remove support for int8 branches, as creating issues downstream in ROOT

+++ PhysicsTools/NanoAOD/python/globals_cff.py
@@ -11,7 +11,7 @@ beamSpotTable = simpleBeamspotFlatTableProducer.clone(
     name = cms.string("BeamSpot"),
     doc = cms.string("offlineBeamSpot, the offline reconstructed beamspot"),
     variables = cms.PSet(
-       type = Var("type()","int8",doc="BeamSpot type (Unknown = -1, Fake = 0, LHC = 1, Tracker = 2)"),
+       type = Var("type()","int16",doc="BeamSpot type (Unknown = -1, Fake = 0, LHC = 1, Tracker = 2)"),
        z = Var("position().z()",float,doc="BeamSpot center, z coordinate (cm)",precision=-1),
        zError = Var("z0Error()",float,doc="Error on BeamSpot center, z coordinate (cm)",precision=-1),
        sigmaZ = Var("sigmaZ()",float,doc="Width of BeamSpot in z (cm)",precision=-1),
